pyproject.toml
[tool.poetry]
name = "pyspark-helpers"
version = "0.1.0"
description = "A collection of tools to help when developing PySpark applications"
authors = ["Jens Peder Meldgaard <jenspederm@gmail.com>"]
readme = "README.md"
packages = [
    {include = "pyspark_helpers"},
]

[tool.poetry.scripts]
psh-schema-from-json = "pyspark_helpers.schema:main"

[tool.poetry.dependencies]
python = "^3.8"
pyspark = "^3.3.2"
delta-spark = "^2.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
